Ensure pep8 standards Add tests to check response object
