+++ requirements.txt
@@ -5,3 +5,4 @@ djangorestframework==3.9.0
 PyJWT==1.7.1
 six==1.12.0 
 psycopg2-binary==2.7.6.1
+pep8==1.7.1
